commented image sequence visualize function

+++ utils.py
@@ -1,16 +1,28 @@
 import cv2
 import os 
 
-def show_image_sequence(image_folder):
+def show_image_sequence(image_folder: str):
+    '''
+    This function takes a folder containing image sequence and visualize them. 
+    Args:
+        image_folder (str): Path to the folder containing image sequence.
+
+    '''
+    if not isinstance(image_folder, str):
+        raise ValueError("Please provide a string as the image_folder argument.")
+    
+    if not os.path.exists(image_folder):
+        print(f"The folder {image_folder} does not exist.")
+        return 
+    
     image_list = sorted(os.listdir(image_folder))
 
-    for image in image_list:
-        # print(image)
+    for image in image_list:        
         img = cv2.imread(os.path.join(image_folder, image))
-        # print(img.shape)
         cv2.imshow('img',img)
         
         if cv2.waitKey(1) & 0xFF == ord('q'):
+            print('Terminated by user')
             cv2.destroyAllWindows()
             break
